win/xpops/parse batch: the Win::close tail was an unqualified close() - infinite self-recursion, now the image's qualified AutoSound::close (0x5ebc80 call 0x62bbf0), the report that surfaced it; vox_fill_colour_table's cascade n was unsigned so every n>=0 was tautological - signed fix stops the runaway fill past the 0x20000 table; parse-string family (ParseStrPlurality/Num/Gender/Buffer 0x9BB570..E8) real objects, parse_num BYTE_EXACT held; X_pop's eight mod-check strings real (.rdata + three pointer slots), the alien gate reads Players[].rule_flags & RFLAG_ALIEN (0x946F58 was &Players[0].rule_flags, 'spying' was a misread); SpyingFactionFlagBytes binding deleted, spying() reads Players directly; boot now parses ALPHAX.txt deep into the rules load and dies at the KNOWN pending popup engine BasePop::start (0x601BF0 thunk, SIGILL) via the BADTECHKEY diagnostic - the tech-name table store is the open question parked with evidence; gate exit 3, 1,952 verified, 0 regressed

+++ src/caviar.cpp
@@ -1129,7 +1129,12 @@ void __cdecl vox_fill_colour_table(void *table, unsigned long value,
         cursor += 2;
     }
     const unsigned long tail_word = count & 1;
-    unsigned long n = count >> 1;  // dword count
+    // SIGNED, and the whole 16/8/4/2/1 cascade hangs on it: the image's
+    // stage guards are jl/jge (0x63AFDD..0x63B066), so the count is meant
+    // to go negative and stop. An unsigned spelling makes every `n >= 0`
+    // tautologically true and the 16-wide loop wraps straight past the
+    // buffer - the page fault this body produced on its first live run.
+    int n = (int)(count >> 1);  // dword count
     if (n != 0) {
         unsigned long *p = (unsigned long *)cursor;
         n -= 16;
